Refactor the components directory style
This is needed since each component should have a lot
of different stuff like:

- test.js
- __mock__ dir
- queries.js
- mutations.js
- __snapshot__ dir
- index.stories.js
