+++ components/index.js
@@ -1,0 +1,5 @@
+import ProductsList from './products-list'
+
+export {
+  ProductsList
+}
